+++ docs/changelog.md
@@ -71,6 +71,10 @@
 
 ## 更新说明详情
 
+1.12.14 add: add stock_hk_profit_forecast_et interface
+
+    1. 新增 stock_hk_profit_forecast_et 接口
+
 1.12.13 fix: fix index_stock_cons_csindex interface
 
     1. 修复 index_stock_cons_csindex 接口
@@ -3321,6 +3325,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.14 add: add stock_hk_profit_forecast_et interface
 
 1.12.13 fix: fix index_stock_cons_csindex interface